// src/app/shared/models/datadaily/DailySaveRequest.ts
export class DailySaveRequest {
    dataMatrixId: number
    baseYearId: number
    yearTerm: number
    monthTerm: number
    usedFlag: string
    dataMatrix: number
     previousPrice: number
     previousAdjustedPrice: number
     currentPrice: number
     currentAdjustedPrice: number
     currentCalculatedPrice:number;
     previousCalculatedPrice:number;
     currencyRate: number
     currencyUnit: number
     noOfCarrierForward: number
     imputationFlag: number
     relativeRatio: number
     link: number
     remark: string;
     requestCommentFlag: string
     answerCommentFlag:string;
     commentNote: string;
     relativeId: number
     price1: number
     price2: number
     price3: number
     price4: number
     price5: number
     price6: number
     price7: number
     price8: number
     price9: number
     price10: number
     price11: number
     price12: number
     price13: number
     price14: number
     price15: number
     price16: number
     price17: number
     price18: number
     price19: number
     price20: number
     price21: number
     price22: number
     price23: number
     price24: number
     price25: number
     price26: number
     price27: number
     price28: number
     price29: number
     price30: number
     price31: number
     rel:number
     priceFlag:String;
     previousCurrencyRate: number
     currentCurrencyRate: number
}